safety-geometry/requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0
datasets>=2.18.0
tqdm>=4.66.0
